requirements.txt
librosa==0.10.1
numpy==1.24.3
python==3.8.17
pytorch==1.12.1
pyyaml==6.0.1
scikit-learn==1.3.0
scipy==1.10.1
tensorboard==2.12.1
tqdm==4.66.1
transformers==4.31.0
matplotlib==3.7.2